requirements.txt
asgiref==3.3.0
certifi==2020.6.20
Django==3.1.2
django-crispy-forms==1.12.0
olefile==0.46
Pillow==8.0.0
pip==21.0.1
psycopg2==2.8.5
pytz==2020.1
setuptools==52.0.0.post20210125
sqlparse==0.4.1
wheel==0.37.0
wincertstore==0.2
